Rename folder for endpoints
