+++ metabase_tools/endpoints/user_endpoint.py
@@ -7,7 +7,7 @@ import logging
 from typing import Any, ClassVar, Optional
 
 from metabase_tools.common import log_call
-from metabase_tools.endpoint.generic_endpoint import Endpoint
+from metabase_tools.endpoints.generic_endpoint import Endpoint
 from metabase_tools.exceptions import InvalidParameters
 from metabase_tools.models.user_model import UserItem
