+++ metabase_tools/metabase.py
@@ -11,10 +11,10 @@ from typing import Any, Optional
 from requests import Response, Session
 from requests.exceptions import RequestException
 
-from metabase_tools.endpoint.cards_endpoint import Cards
-from metabase_tools.endpoint.collection_endpoint import Collections
-from metabase_tools.endpoint.database_endpoint import Databases
-from metabase_tools.endpoint.user_endpoint import Users
+from metabase_tools.endpoints.cards_endpoint import Cards
+from metabase_tools.endpoints.collection_endpoint import Collections
+from metabase_tools.endpoints.database_endpoint import Databases
+from metabase_tools.endpoints.user_endpoint import Users
 from metabase_tools.exceptions import (
     AuthenticationFailure,
     InvalidDataReceived,
